+++ frontend/CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add "Last" button for pagination based on cursors.
+
+### Changed
+
 - Move query `useBlockSpecialEventsQuery` to new API.
+
+### Fixed
+
 - Rename `ContractUpgraded.from` and `ContractUpgraded.to` in query `useTransactionQuery` to avoid error of fields which cannot be merged, as these were clashing with `Transferred.from` and `Transferred.to`.
 
 ## [1.6.2] - 2025-01-19
@@ -13,7 +22,7 @@ All notable changes to this project will be documented in this file.
 
 - Allowed queries `useBlockListQuery`, `useBlockQuery`, `useChartBlockMetrics`, `useContractQuery`, `useContractsListQuery`, `useTokenQuery`, `useTokenListQuery` and `useModuleQuery` to use a separate API.
 
-### Fix
+### Fixed
 
 - Remove unused fields in query for cis2 token events.
 - Remove unused sub-query for account transactions in the accounts page.